fix static visualizations dashcard settings merging
