+++ frontend/src/metabase/static-viz/components/ScalarChart/ScalarChart.stories.tsx
@@ -38,6 +38,5 @@ const renderingContext: RenderingContext = {
 export const Default = Template.bind({});
 Default.args = {
   rawSeries: data.twoScalars as any,
-  dashcardSettings: {},
   renderingContext,
 };
