+++ frontend/src/metabase/static-viz/components/ScatterPlot/ScatterPlot.tsx
@@ -21,7 +21,6 @@ const LEGEND_PADDING = 8;
 
 export function ScatterPlot({
   rawSeries,
-  dashcardSettings,
   renderingContext,
   width = WIDTH,
   height = HEIGHT,
@@ -31,7 +30,6 @@ export function ScatterPlot({
 
   const computedVisualizationSettings = computeStaticComboChartSettings(
     rawSeries,
-    dashcardSettings,
     renderingContext,
   );
