+++ frontend/src/metabase/static-viz/components/StaticVisualization/types.ts
@@ -1,9 +1,8 @@
 import type { RenderingContext } from "metabase/visualizations/types";
-import type { RawSeries, VisualizationSettings } from "metabase-types/api";
+import type { RawSeries } from "metabase-types/api";
 
 export interface StaticChartProps {
   rawSeries: RawSeries;
-  dashcardSettings: VisualizationSettings;
   renderingContext: RenderingContext;
   width?: number;
   height?: number;
