+++ frontend/src/metabase/static-viz/index.js
@@ -2,6 +2,7 @@ import "fast-text-encoding";
 
 import { setPlatformAPI } from "echarts/core";
 import ReactDOMServer from "react-dom/server";
+
 import "metabase/lib/dayjs";
 
 import { StaticVisualization } from "metabase/static-viz/components/StaticVisualization";
@@ -12,6 +13,8 @@ import {
   measureTextHeight,
   measureTextWidth,
 } from "metabase/static-viz/lib/text";
+import { extractRemappings } from "metabase/visualizations";
+import { extendCardWithDashcardSettings } from "metabase/visualizations/lib/settings/typed-utils";
 import { DEFAULT_VISUALIZATION_THEME } from "metabase/visualizations/shared/utils/theme";
 
 import { LegacyStaticChart } from "./containers/LegacyStaticChart";
@@ -32,6 +35,21 @@ export function LegacyRenderChart(type, options) {
   );
 }
 
+// Dashcard settings should be merged with the first card settings
+// Replicates the logic from frontend/src/metabase/dashboard/components/DashCard/DashCard.tsx
+function getRawSeriesWithDashcardSettings(rawSeries, dashcardSettings) {
+  return rawSeries.map((series, index) => {
+    const isMainCard = index === 0;
+    if (isMainCard) {
+      return {
+        ...series,
+        card: extendCardWithDashcardSettings(series.card, dashcardSettings),
+      };
+    }
+    return series;
+  });
+}
+
 export function RenderChart(rawSeries, dashcardSettings, colors) {
   const getColor = createColorGetter(colors);
   const renderingContext = {
@@ -44,13 +62,18 @@ export function RenderChart(rawSeries, dashcardSettings, colors) {
     theme: DEFAULT_VISUALIZATION_THEME,
   };
 
-  const props = {
+  const rawSeriesWithDashcardSettings = getRawSeriesWithDashcardSettings(
     rawSeries,
     dashcardSettings,
-    renderingContext,
-  };
+  );
+  const rawSeriesWithRemappings = extractRemappings(
+    rawSeriesWithDashcardSettings,
+  );
 
   return ReactDOMServer.renderToStaticMarkup(
-    <StaticVisualization {...props} />,
+    <StaticVisualization
+      rawSeries={rawSeriesWithRemappings}
+      renderingContext={renderingContext}
+    />,
   );
 }
